Removed old async dispose helper
